Redesigned position reading and login modules

+++ ServerInSpace/ServerInSpace-ejb/src/java/com/objects/Spaceship.java
@@ -9,10 +9,10 @@ public class Spaceship {
     
     private int id;
 
-    private float x, y;
+    private double x, y;
     private int direction;
 
-    public Spaceship(int id, float x, float y, int direction) {
+    public Spaceship(int id, double x, double y, int direction) {
         this.id = id;
         this.x = x;
         this.y = y;
@@ -27,19 +27,19 @@ public class Spaceship {
         this.id = id;
     }
 
-    public float getX() {
+    public double getX() {
         return x;
     }
  
-    public void setX(float x) {
+    public void setX(double x) {
         this.x = x;
     }
  
-    public float getY() {
+    public double getY() {
         return y;
     }
  
-    public void setY(float y) {
+    public void setY(double y) {
         this.y = y;
     }
  
@@ -51,7 +51,7 @@ public class Spaceship {
         this.direction = direction;
     }
     
-    public void update(float x, float y, int direction){
+    public void update(double x, double y, int direction){
         this.x = x;
         this.y = y;
         this.direction = direction;
